3 more green, 45 red left
